organize project files, tidy up code
